services/streamlit-ui/requirements.txt
streamlit==1.28.2
pymongo==4.6.0
pandas==2.0.3
plotly==5.18.0
python-dotenv==1.0.0